Move ticket booking to the right project
