+++ GreenAfrica-API/Data/FlightInfo.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using GreenAfrica.DataAccess.Models;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
