+++ GreenAfrica-API/Repo/ITicketBookingRepository.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
+using GreenAfrica.DataAccess.Models;
 
 namespace GreenAfrica_API.Repo
 {
